pyproject.toml
# Copyright 2024 Google LLC
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#      https://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

[project]
name = "ml_goodput_measurement"
version = "0.0.5"
authors = [
  { name="Cloud TPU Team", email="cloud-tpu-eng@google.com" },
]
description = "Package to retrieve Goodput of jobs running on Cloud TPU."
readme = "README.md"
requires-python = ">=3.8"
license = {text = "Apache-2.0"}
classifiers = [
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
keywords = []

# pip dependencies installed with `pip install -e .`
dependencies = [
  "google-cloud-logging>=3.5.0",
  "google-cloud-monitoring>=2.20.0",
  "numpy",
  "scipy",
  "tensorboardx",
]

[project.urls]
"Homepage" = "https://github.com/AI-Hypercomputer/ml-goodput-measurement"
"Bug Tracker" = "https://github.com/AI-Hypercomputer/ml-goodput-measurement/issues"

[build-system]
# Build system specify which backend is used to build/install the project
requires = ["flit_core >=3.8,<4"]
build-backend = "flit_core.buildapi"

[tool.flit.sdist]
# Flit specific options (files to exclude from the PyPI package)
exclude = [
  # Do not release tests files on PyPI
  "tests/*_test.py",
]